+++ upcloud/provider.go
@@ -101,7 +101,6 @@ func Provider() *schema.Provider {
 			"upcloud_tags":               tag.DataSourceTags(),
 			"upcloud_storage":            storage.DataSourceStorage(),
 			"upcloud_kubernetes_cluster": kubernetes.DataSourceCluster(),
-			"upcloud_kubernetes_plan":    kubernetes.DataSourcePlan(),
 		},
 
 		ConfigureContextFunc: providerConfigure,